Payement_app/Extract_Strore/Extract_Strore_requirements.txt
langchain==1.0.1
langchain_core==1.0.0
langchain_google_genai==3.0.0
langchain_postgres==0.0.16
numpy==2.3.4
protobuf==6.33.0
python_pptx==1.0.2
SQLAlchemy==2.0.44
